+++ frex/stores/exceptions/custom_exception.py
